feat: 내 지도에서 프로필 이미지 -> 공유 버튼으로 변경  (#210)
* chore: 공유 아이콘 변경

* feat: 공유 버튼 클릭 시, 공유 바텀 시트 보이도록 수정

* feat: 내 지도에서 프로필 -> 공유 버튼으로 변경

* chore: 사용하지 않는 ref 삭제

* chore: 하단 탭 추가로 인해 사용하지 않는 컴포넌트 삭제

* fix: 사용하지 않는 props 삭제
