+++ src/features/home/components/lifeMap/PrivateLifeMap.tsx
@@ -1,5 +1,3 @@
-import { useRef } from 'react';
-
 import { BottomNavigation } from '@/components/molecules/bottomNavigation';
 import { useGetMemberData } from '@/hooks/reactQuery/auth';
 import { useGetGoals } from '@/hooks/reactQuery/goal/useGetGoals';
@@ -10,11 +8,9 @@ export const PrivateLifeMap = () => {
   const { data: memberData } = useGetMemberData();
   const { data: privateGoals } = useGetGoals();
 
-  const downloadSectionRef = useRef<HTMLElement>(null);
-
   return (
     <>
-      <LifeMapContent goalsData={privateGoals} memberData={memberData} downloadSectionRef={downloadSectionRef} />
+      <LifeMapContent goalsData={privateGoals} memberData={memberData} />
       <BottomNavigation />
     </>
   );
