+++ src/features/home/components/lifeMap/PublicLifeMap.tsx
@@ -6,14 +6,12 @@ import { useGetPublicGoals } from '@/hooks/reactQuery/goal/useGetPublicGoals';
 
 import { LifeMapContent } from './LifeMapContent';
 
-const CHEER_ANIMATION_INTERVAL = 5400;
-
 export const PublicLifeMap = ({ username }: { username: string }) => {
   const { data: publicGoals } = useGetPublicGoals({ username });
 
   return (
     <>
-      <LifeMapContent isPublic goalsData={publicGoals} memberData={publicGoals?.user} />
+      <LifeMapContent goalsData={publicGoals} memberData={publicGoals?.user} />
       <PublicLifeMapBottomArea username={username} />
     </>
   );
